src/dify/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-research"
version = "0.1.0"
description = "Gemini Research Tools for Dify"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dify-plugin>=0.7.1",
    "google-genai>=1.57.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
